+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+- add support for `--dart-define-from-file` argument
+- dependencies update
+- add mode compile
+- add suppor request fromData and x-www-form-urlencoded
+
 ## 0.0.1
 
 - initial implementation of classes